OFZ-26 refactor: menuStore -> useMenuStore 이름 변경
